JWT extended

+++ resources/item.py
@@ -1,5 +1,11 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required, 
+    get_jwt_claims, 
+    jwt_optional, 
+    get_jwt_identity, 
+    fresh_jwt_required
+    )
 from models.item import ItemModel
 
 class Item(Resource):
@@ -15,14 +21,14 @@ class Item(Resource):
         help="Every item needs a store id"
         )
         
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json(), 200
         return {'message': 'Item not found'}, 404
         
-        
+    @fresh_jwt_required  
     def post(self, name):
         # force=True - for json format
         # silent=True - returns None insetad of error
@@ -39,7 +45,11 @@ class Item(Resource):
             return {'message': 'An error occurred'}, 500
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'You don\'t have permissions to do this operation.'}
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -60,5 +70,11 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}
+        return {'items': [item['name'] for item in items],
+                'message': 'More data available after login'}
